# Conflicts:
#	apps/api-gateway/src/resolvers/rest.ts
#	apps/indexer/ponder.schema.ts
#	apps/indexer/src/api/controller/index.ts
#	apps/indexer/src/api/index.ts
#	apps/indexer/src/eventHandlers/delegation.ts
#	apps/indexer/src/eventHandlers/transfer.ts
#	apps/indexer/src/indexer/arb/erc20.ts
#	apps/indexer/src/indexer/ens/erc20.ts
#	apps/indexer/src/indexer/op/erc20.ts
#	apps/indexer/src/indexer/uni/erc20.ts

+++ apps/indexer/src/indexer/arb/erc20.ts
@@ -26,7 +26,7 @@ export function ARBTokenIndexer(address: Address, decimals: number) {
       timestamp: event.block.timestamp,
       transactionFrom: event.transaction.from,
       transactionTo: event.transaction.to,
-      logIndex: BigInt(event.log.logIndex),
+      logIndex: event.log.logIndex,
     });
   });
 }
